Allow to set the minimum macOS version.
When building the macOS version, allow to pass a --with-min-osx-version
option to the configure script to explicitly set the minimal macOS
version the launcher will be able to run on.

If that option is not set, the minimal version required to run the
launcher will depend on the default configuration of the build
environment. On macOS, the XCode command line tools seem to default to
target the same version as the version they are currently running on.