Compile the Windows version as a GUI application.
Use the -mwindows flag when building the Windows version, so that the
launcher behaves as a graphical application -- meaning a console window
will not be opened if the launcher is started from the desktop (e.g. by
double-clicking on the program).

If the program is started from a terminal however, whoever started it
that is probably interested in seeing the messages in the console for
debugging purposes, so we attach the console from the parent process
(the shell).