Support the legacy Protege.l4j.ini file.
Current versions of Protégé for Windows can be configured through a
Protege.l4j.ini file located alongside the launcher. The new jvm.conf
format is to be preferred, but for backwards compatibility, if there is
no jvm.conf file but a Protege.l4j.ini file is found, we get the Java
options from that file.