Support legacy options even if a jvm.conf file is present.
On macOS and Windows, if we found a jvm.conf file but that file didn't
contain any extra options, we still look for options using the "legacy"
mechanisms (Info.plist for macOS, Protege.l4j.ini for Windows), instead
of only looking there when a jvm.conf is absent.

This makes the legacy mechanisms available even in the default
configuration, with the empty jvm.conf file shipped as part of Protégé's
distribution.