Try setting a default max heap setting.
If the user does not explicitly set a value for the max heap size
(Java's -Xmx option), the default value is set by the Java virtual
machine. With most JVMs, it is set to ~1GB, which is likely to be too
low for many Protégé users.

We thus try to determine a better default value based on the total
amount of physical memory available:

- less than 4 GB: we stick to the JVM's default value;
- between 4 and 8 GB: we use half of the physical memory;
- between 8 and 16 GB: we use two-thirds of the physical memory;
- more than 16 GB: we use three-quarters of the physical memory.