Allow use of '%' when setting memory options.
If the numerical value of a memory option is followed by a '%' character
(instead of one of "kKmMgG"), set the value to the corresponding
fraction of the total physical memory available on the system.

That is, `max_heap_size=50%` would set the max heap size to 4192 MB on a
machine with 8192 MB of RAM.

This is only available when using the modern configuration file
(jvm.conf), *not* when using legacy mechanisms (Info.plist on macOS and
Protege.l4j.ini on Windows).