Pass command line arguments to the OSGI launcher.
The OSGI launcher can deal with command line arguments and make sure
the application is aware of them after it has started, so if we did get
arguments in the C launcher, we can pass them to the OSGI launcher.

This allows the Windows version of Protégé to start when users
double-click on an ontology file (provided the file format has been
associated with Protégé); same on GNU/Linux.