Try auto UI scaling on GNU/Linux.
On Windows and MacOS, the Java virtual machine somehow automatically
detects whether it is running on a macine with a "HiDPI" display, and
automatically scales Swing widgets accordingly.

This does not seem to work automatically on GNU/Linux though, so here we
try (1) detecting whether we have a "HiDPI" display (defined, for our
purpose, as a display where the resolution is 144 DPI or more) and (2)
if so, append a `-Dsun.java2d.uiScale=X` option for the JVM, where _X_
is the scaling factor.

This feature is automatically disabled if the user has already set a
`sun.java2d.uiScale` option in the configuration file.