Add an explicit option to enable/disable auto UI scaling.
Add a `ui_auto_scaling=<yes|no>` option in the jvm.conf file to
explicitly enable or disable the automatic UI scaling feature on
GNU/Linux.

The configure-time option now serves, not to enable or disable the
feature (it is always compiled in), but to make it enabled by default or
not.